Did a lot of changes, InputManager and Completed two stupid IInputStates, also trying to run the unittest and test those states. Also did constant correctness all the LevelManager stuff

+++ src/GameUnits/GameObject.cpp
@@ -14,7 +14,8 @@ CollisionManager *GameObject::collMng = 0;
 
 
 ////////////////////////////////////////////////////////////////////////////////
-void GameObject::configCollObj(float w, float h, float mF, float mG)
+void
+GameObject::configCollObj(float w, float h, uint32 mF, uint32 mG)
 {
 	mCollObject.bb.setSize(w, h);
 	mCollObject.maskFlag = mF;
